+++ test/console.js
@@ -2,8 +2,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-const assert = require('assert');
 const util = require('util');
+
+const assert = require('insist');
 
 const intel = require('../');
 const consoleUtil = require('./util/console');