+++ test/handlers.js
@@ -2,11 +2,12 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-const assert = require('assert');
 const EventEmitter = require('events').EventEmitter;
 const fs = require('fs');
 const os = require('os');
 const path = require('path');
+
+const assert = require('insist');
 
 const intel = require('../');